refactor: split screen_reader_summary into focused utility modules
