+++ packages/charts/src/components/accessibility/screen_reader_summary.tsx
@@ -9,34 +9,12 @@
 import React, { memo } from 'react';
 import { connect } from 'react-redux';
 
+import { createAxisDescriptions } from './axis_summary_utils';
+import { createChartTypeDescription } from './chart_summary_utils';
 import { ScreenReaderDescription } from './description';
+import { createGoalChartDescription } from './goal_chart_summary_utils';
 import { ScreenReaderLabel } from './label';
-import { ChartType } from '../../chart_types';
-import type { GoalChartData, GoalChartLabels } from '../../chart_types/goal_chart/state/selectors/get_goal_chart_data';
-import {
-  getGoalChartDataSelector,
-  getGoalChartLabelsSelector,
-} from '../../chart_types/goal_chart/state/selectors/get_goal_chart_data';
-import { computeSeriesDomainsSelector } from '../../chart_types/xy_chart/state/selectors/compute_series_domains';
-import { getAxisSpecsSelector, getSeriesSpecsSelector } from '../../chart_types/xy_chart/state/selectors/get_specs';
-import type { SeriesDomainsAndData } from '../../chart_types/xy_chart/state/utils/types';
-import type { BasicSeriesSpec, AxisSpec } from '../../chart_types/xy_chart/utils/specs';
-import type { GlobalChartState } from '../../state/chart_state';
-import type { A11ySettings } from '../../state/selectors/get_accessibility_config';
-import { DEFAULT_A11Y_SETTINGS, getA11ySettingsSelector } from '../../state/selectors/get_accessibility_config';
-import { getInternalChartStateSelector } from '../../state/selectors/get_internal_chart_state';
-import { getInternalIsInitializedSelector, InitStatus } from '../../state/selectors/get_internal_is_intialized';
-
-interface ScreenReaderSummaryStateProps {
-  a11ySettings: A11ySettings;
-  chartTypeDescription: string;
-  goalChartData?: GoalChartData;
-  goalChartLabels?: GoalChartLabels;
-  seriesSpecs?: BasicSeriesSpec[];
-  axisSpecs?: AxisSpec[];
-  seriesDomains?: SeriesDomainsAndData;
-  chartType: ChartType | null;
-}
+import { mapStateToProps, type ScreenReaderSummaryStateProps } from './screen_reader_summary_connector';
 
 const ScreenReaderSummaryComponent = ({
   a11ySettings,
@@ -47,166 +25,25 @@ const ScreenReaderSummaryComponent = ({
   axisSpecs,
   seriesDomains,
 }: ScreenReaderSummaryStateProps) => {
-  // Create a consolidated summary text for better screen reader experience
   const createConsolidatedSummary = () => {
     const parts: string[] = [];
 
     // Chart type and series information
-    if (chartTypeDescription && seriesSpecs && seriesDomains?.formattedDataSeries) {
-      const seriesTypes = new Set<string>();
-      seriesSpecs.forEach((spec) => seriesTypes.add(spec.seriesType));
-
-      // Get actual rendered series count (after split processing)
-      const actualSeriesCount = seriesDomains.formattedDataSeries.length;
-      const renderedSeries = seriesDomains.formattedDataSeries;
-
-      // Get series names from rendered series (which includes split series)
-      const seriesNames = renderedSeries
-        .map((series) => {
-          // Extract series name from the series key or use the spec name
-          const spec = seriesSpecs.find((s) => s.id === series.specId);
-          if (spec && typeof spec.name === 'string') return spec.name;
-          if (series.splitAccessors && series.splitAccessors.size > 0) {
-            // For split series, use the split accessor values as the name
-            const splitValues = Array.from(series.splitAccessors.values());
-            return splitValues.join(' ');
-          }
-          return `Series ${series.specId}`;
-        })
-        .filter((name) => name);
-
-      // Check if any series are stacked
-      const hasStackedSeries = seriesSpecs.some(
-        (spec) => 'stackAccessors' in spec && spec.stackAccessors && spec.stackAccessors.length > 0,
-      );
-
-      if (seriesTypes.size === 1) {
-        const seriesType = Array.from(seriesTypes)[0];
-
-        // Check for percentage stacking
-        const hasPercentageStacking = seriesSpecs.some(
-          (spec) => 'stackMode' in spec && spec.stackMode === 'percentage',
-        );
-
-        const stackPrefix = hasStackedSeries ? (hasPercentageStacking ? 'percentage stacked' : 'stacked') : '';
-
-        if (actualSeriesCount === 1) {
-          parts.push(`${stackPrefix ? `${stackPrefix} ` : ''}${seriesType} chart`);
-        } else {
-          const chartTypeDescriptionStackChecked = `${stackPrefix ? `${stackPrefix} ` : ''}${seriesType} chart`;
-          const countDescription = `with ${actualSeriesCount} ${seriesType}s`;
-
-          const description = `${chartTypeDescriptionStackChecked} ${countDescription}`;
-          if (seriesNames.length > 0 && seriesNames.length <= 5) {
-            parts.push(`${description}: ${seriesNames.join(', ')}`);
-          } else {
-            parts.push(description);
-          }
-        }
-      } else {
-        const stackPrefix = hasStackedSeries ? 'stacked mixed' : 'mixed';
-        parts.push(`${stackPrefix} chart: ${Array.from(seriesTypes).join(' and ')} chart`);
-      }
-    } else if (chartTypeDescription) {
-      parts.push(chartTypeDescription);
+    const chartDescription = createChartTypeDescription(chartTypeDescription, seriesSpecs, seriesDomains);
+    if (chartDescription) {
+      parts.push(chartDescription);
     }
 
     // Goal chart specific data
-    const validGoalChart =
-      chartTypeDescription === 'goal chart' ||
-      chartTypeDescription === 'horizontalBullet chart' ||
-      chartTypeDescription === 'verticalBullet chart';
-
-    if (validGoalChart && goalChartData && !isNaN(goalChartData.maximum)) {
-      parts.push(
-        `Minimum: ${goalChartData.minimum}, Maximum: ${goalChartData.maximum}, Target: ${goalChartData.target}, Value: ${goalChartData.value}`,
-      );
+    const goalDescription = createGoalChartDescription(chartTypeDescription, goalChartData);
+    if (goalDescription) {
+      parts.push(goalDescription);
     }
 
     // Axis descriptions
     if (axisSpecs && axisSpecs.length > 0 && seriesDomains) {
-      const { xDomain, yDomains } = seriesDomains;
-
-      // X axis description
-      const xAxisSpecs = axisSpecs.filter((spec) => spec.position === 'bottom' || spec.position === 'top');
-      if (xAxisSpecs.length > 0 && xDomain) {
-        const xAxisSpec = xAxisSpecs[0]!;
-        const axisTitle = xAxisSpec.title || 'X';
-        let rangeDescription = '';
-
-        switch (xDomain.type) {
-          case 'ordinal':
-            rangeDescription = `with ${xDomain.domain.length} categories`;
-            break;
-          case 'time': {
-            const minTime = new Date(xDomain.domain[0]);
-            const maxTime = new Date(xDomain.domain[1]);
-            const timeDiff = maxTime.getTime() - minTime.getTime();
-
-            // Choose granularity based on time span
-            const formatTime = (date: Date) => {
-              if (timeDiff < 24 * 60 * 60 * 1000) {
-                // Less than 1 day
-                return date.toLocaleString('en-US', {
-                  month: 'short',
-                  day: 'numeric',
-                  hour: 'numeric',
-                  minute: '2-digit',
-                });
-              } else if (timeDiff < 7 * 24 * 60 * 60 * 1000) {
-                // Less than 1 week
-                return date.toLocaleDateString('en-US', {
-                  month: 'short',
-                  day: 'numeric',
-                  hour: 'numeric',
-                });
-              } else if (timeDiff < 365 * 24 * 60 * 60 * 1000) {
-                // Less than 1 year
-                return date.toLocaleDateString('en-US', {
-                  year: 'numeric',
-                  month: 'short',
-                  day: 'numeric',
-                });
-              } else {
-                // 1 year or more
-                return date.toLocaleDateString('en-US', {
-                  year: 'numeric',
-                  month: 'long',
-                });
-              }
-            };
-            rangeDescription = `from ${formatTime(minTime)} to ${formatTime(maxTime)}`;
-            break;
-          }
-          case 'linear':
-            const [min, max] = xDomain.domain;
-            rangeDescription = `from ${min} to ${max}`;
-            break;
-        }
-
-        parts.push(`X axis displays ${axisTitle}${rangeDescription ? ` ${rangeDescription}` : ''}`);
-      }
-
-      // Y axes description
-      const yAxisSpecs = axisSpecs.filter((spec) => spec.position === 'left' || spec.position === 'right');
-      if (yAxisSpecs.length > 0 && yDomains.length > 0) {
-        const primaryYDomain = yDomains[0];
-        const yAxisSpec = yAxisSpecs[0]!;
-        const axisTitle = yAxisSpec.title || 'Values';
-
-        let rangeDescription = '';
-        if (primaryYDomain && primaryYDomain.domain.length >= 2) {
-          const [min, max] = primaryYDomain.domain;
-          rangeDescription = `, ranging from ${min} to ${max}`;
-        }
-
-        const yAxisDescription =
-          yAxisSpecs.length === 1
-            ? `Y axis displays ${axisTitle}${rangeDescription}`
-            : `${yAxisSpecs.length} Y axes${rangeDescription}`;
-
-        parts.push(yAxisDescription);
-      }
+      const axisDescriptions = createAxisDescriptions(axisSpecs, seriesDomains);
+      parts.push(...axisDescriptions);
     }
 
     return `${parts.join('. ')}.`;
@@ -223,49 +60,6 @@ const ScreenReaderSummaryComponent = ({
   );
 };
 
-const DEFAULT_SCREEN_READER_SUMMARY = {
-  a11ySettings: DEFAULT_A11Y_SETTINGS,
-  chartTypeDescription: '',
-  goalChartData: undefined,
-  seriesSpecs: undefined,
-  axisSpecs: undefined,
-  seriesDomains: undefined,
-  chartType: null,
-};
-
-const mapStateToProps = (state: GlobalChartState): ScreenReaderSummaryStateProps => {
-  const internalChartState = getInternalChartStateSelector(state);
-  if (internalChartState === null || getInternalIsInitializedSelector(state) !== InitStatus.Initialized) {
-    return DEFAULT_SCREEN_READER_SUMMARY;
-  }
-
-  // Get XY chart data when available
-  let seriesSpecs: BasicSeriesSpec[] | undefined;
-  let axisSpecs: AxisSpec[] | undefined;
-  let seriesDomains: SeriesDomainsAndData | undefined;
-
-  try {
-    // Only get XY chart selectors if this is an XY chart
-    if (state.chartType === ChartType.XYAxis) {
-      seriesSpecs = getSeriesSpecsSelector(state);
-      axisSpecs = getAxisSpecsSelector(state);
-      seriesDomains = computeSeriesDomainsSelector(state);
-    }
-  } catch {
-    // Silently handle any selector errors - these selectors may not apply to all chart types
-  }
-
-  return {
-    chartTypeDescription: internalChartState.getChartTypeDescription(state),
-    a11ySettings: getA11ySettingsSelector(state),
-    goalChartData: getGoalChartDataSelector(state),
-    goalChartLabels: getGoalChartLabelsSelector(state),
-    seriesSpecs,
-    axisSpecs,
-    seriesDomains,
-    chartType: state.chartType,
-  };
-};
 
 /** @internal */
 export const ScreenReaderSummary = memo(connect(mapStateToProps)(ScreenReaderSummaryComponent));
